Refactor `.pipe()` logic
