+++ package.json
@@ -52,6 +52,7 @@
 		"ava": "^6.1.3",
 		"get-node": "^15.0.1",
 		"path-key": "^4.0.0",
+		"tempy": "^3.1.0",
 		"typescript": "^5.5.4",
 		"xo": "^0.59.3",
 		"yoctocolors": "^2.1.1"
