+++ pipe.js
@@ -1,15 +1,6 @@
 import {pipeline} from 'node:stream/promises';
-import {normalizeArguments, getCommand, spawnProcess} from './spawn.js';
-import {addPromiseMethods} from './iterable.js';
 
-export const pipe = (previous, file, commandArguments, options) => {
-	[commandArguments, options] = normalizeArguments(commandArguments, options);
-	const command = `${previous.context.command} | ${getCommand(file, commandArguments)}`;
-	const context = {...previous.context, command};
-	const resultPromise = spawnProcess(file, commandArguments, {...options, stdin: 'pipe'}, context);
-	const mergedPromise = Object.assign(runProcesses([previous.resultPromise, resultPromise]), resultPromise);
-	return addPromiseMethods(mergedPromise, context);
-};
+export const handlePipe = (previous, resultPromise) => Object.assign(runProcesses([previous.resultPromise, resultPromise]), resultPromise);
 
 const runProcesses = async resultPromises => {
 	// Ensure both subprocesses have exited before resolving, and that we handle errors from both
@@ -25,19 +16,26 @@ const runProcesses = async resultPromises => {
 };
 
 const pipeStreams = async resultPromises => {
-	const {stdin} = await resultPromises[1].nodeChildProcess;
-
 	try {
-		const {stdout} = await resultPromises[0].nodeChildProcess;
+		const [{stdout}, {stdin}] = await Promise.all(resultPromises.map(({nodeChildProcess}) => nodeChildProcess));
+		if (stdin === null) {
+			throw new Error('The "stdin" option must be set on the first "nanoSpawn()" call in the pipeline.');
+		}
+
 		if (stdout === null) {
-			throw new Error('The "stdout" option cannot be combined with ".pipe()".');
+			throw new Error('The "stdout" option must be set on the last "nanoSpawn()" call in the pipeline.');
 		}
 
 		// Do not `await` nor handle stream errors since this is already done by each subprocess
 		// eslint-disable-next-line promise/prefer-await-to-then
 		pipeline(stdout, stdin).catch(() => {});
 	} catch (error) {
-		stdin.end();
+		await Promise.allSettled(resultPromises.map(({nodeChildProcess}) => closeStdin(nodeChildProcess)));
 		throw error;
 	}
+};
+
+const closeStdin = async nodeChildProcess => {
+	const {stdin} = await nodeChildProcess;
+	stdin.end();
 };
